:sparkles: Add pipeline processor support
Added handling in the CommandProcessorBase and pipeline handler for
disposal of script commands, both compiled and interpreted.

NOTE: These changes invoke command disposal methods immediately after
the command's end{} or EndProcessing() call completes, rather than
waiting for the disposal of the whole pipeline to dispose commands.
This changes current behaviour of IDisposable cmdlets, but likely not in
a significant way. If anything, it should improve the memory consumption
of the pipeline handler for longer pipeline command sequences.

 - Added support for setting breakpoints in a Dispose{} block.
 - Added `-DisposeScript` parameter to ForEach-Object.