:bug: Fix NRE in UnixComputer Dispose method
I'm not sure why this test was not throwing errors on every run.
The test behaviour appears to skip the creation of the process object,
which should mean that _process is always null when Dispose() is called
during a test run. I would guess that the error may surface elsewhere
depending on when the pipeline gets disposed.

With the preceding changes, Dispose() is much more consistently called.
As a result this NRE becomes more problematic and must be handled.

Added a check for null before trying to call Dispose() on _process.