:wrench: Dispose commands when complete
Pipeline processor now calls Dispose() on commands after DoComplete().
This is preferable to the previous implementation as it keeps the
pipeline processor in control of the command disposal, but ensures
commands are disposed once their processing is completed, so that their
resources are disposed of when they're done processing.

Also, fixed a couple complaints from Codacy about possible NREs.